# docs/contracts/party-registry.openapi.yaml
openapi: 3.0.3
info:
  title: Party Management API
  version: 1.0.0
  description: |
    Contrato completo del microservicio de Parties usando envelope ApiResponse.

    Todas las respuestas siguen el formato:
    - status
    - code
    - data (cuando aplique)

    Para listados que requieren paginación se incluyen además:
    - nextCursor
    - prevCursor
    - totalElements
    - totalPages
    - numberOfElements

    No se exponen endpoints externos para party_outbox_events.
servers:
  - url: /
...
paths:
...
  /v1/parties:
    get:
      tags:
        - parties
      operationId: listParties
      summary: List tenant-owned Party summaries
      description: |
        Returns only the requesting tenant's summaries, including both types and all four record
        statuses (including ARCHIVED) when the corresponding filter is absent. Stored text, audit
        values, and versions are preserved. No If-Match or Idempotency-Key is required.
        Accepts only type, recordStatus, displayNameStartsWith, displayNameContains, createdFrom,
        createdTo, cursor, and limit, each at most once. Unknown or repeated parameters and invalid
        values return 400 bad-request; malformed filters are never silently ignored.
        All effective filters use AND semantics. Enums match exactly. Name comparison applies
        strip().toUpperCase(Locale.ROOT) to the stored name and supplied filter without rewriting
        stored values. Prefix/substring matching is literal: %, _, accents, punctuation, and interior
        whitespace retain their meaning. Empty or whitespace-only name filters are absent. Each
        decoded name filter is limited to 300 Unicode code points before normalization.
        Creation bounds are inclusive offset-qualified date-times compared as instants. Equivalent
        offsets select the same boundary; equal bounds are permitted, inverted bounds are rejected.
        Results are ordered by createdAt DESC, then canonical unsigned partyId DESC on every page,
        including backward navigation. The default limit is 50; supplied limits must be decimal
        integers from 1 through 200. Blank, fractional, negative, zero, or oversized limits are invalid.
        Opaque cursors bind the tenant, all effective filters, and page size. Canonically equivalent
        name filters and equivalent date-time offsets have the same scope. Omission requests the
        first page; empty, malformed, altered, or scope-mismatched cursors return 400 bad-request.
        Forward/backward continuation is stable on unchanged data. Separate requests see current
        data, not a retained historical snapshot; an authenticated boundary need not still match a row.
        Each response's summaries, exact totalElements, totalPages, and neighbor cursors describe
        one consistent view. Totals count all matches before continuation, totalPages is the ceiling
        of totalElements divided by the effective limit, and numberOfElements is the array length.
        Both cursor properties are always present and use explicit null for unavailable directions.
        When all matches disappear, data is [], both cursors are null, and all three counts are zero.
        An empty continuation page may still offer navigation relative to its authenticated boundary.
        Accepted Process-Id values are echoed unchanged on success and failure. Framework 404/405
        and unexpected 500 failures use the shared status/code-only error envelope.
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/ProcessId'
        - $ref: '#/components/parameters/UserId'
        - name: type
          in: query
          required: false
          description: Exact immutable Party type; omission includes both types.
          schema:
            $ref: '#/components/schemas/PartyType'
        - name: recordStatus
          in: query
          required: false
          description: Exact lifecycle status; omission includes DRAFT, ACTIVE, INACTIVE, and ARCHIVED.
          schema:
            $ref: '#/components/schemas/PartyRecordStatus'
        - name: displayNameStartsWith
          in: query
          required: false
          allowEmptyValue: true
          description: Literal canonical prefix; at most 300 decoded Unicode code points before normalization.
          schema:
            type: string
            maxLength: 300
        - name: displayNameContains
          in: query
          required: false
          allowEmptyValue: true
          description: Literal canonical substring; at most 300 decoded Unicode code points before normalization.
          schema:
            type: string
            maxLength: 300
        - name: createdFrom
          in: query
          required: false
          description: Inclusive creation instant; an explicit UTC offset is required.
          schema:
            type: string
            format: date-time
        - name: createdTo
          in: query
          required: false
          description: Inclusive creation instant; an explicit UTC offset is required and the value cannot precede createdFrom.
          schema:
            type: string
            format: date-time
        - $ref: '#/components/parameters/PartyCursorQuery'
        - $ref: '#/components/parameters/PartyLimitQuery'
      responses:
        '200':
          description: Tenant-qualified summaries and exact, snapshot-consistent pagination metadata
          headers:
            Process-Id:
              $ref: '#/components/headers/ProcessIdEcho'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartyCollectionApiResponse'
              examples:
                empty:
                  summary: No matching records
                  value:
                    status: 200
                    code: successful
                    data: []
                    nextCursor: null
                    prevCursor: null
                    totalElements: 0
                    totalPages: 0
                    numberOfElements: 0
                singlePage:
                  summary: One retained Party and no other matching records
                  value:
                    status: 200
                    code: successful
                    data:
                      - partyId: 0198ce2a-7b7d-7ab4-a5cf-4d4d7db89ab1
                        type: NATURAL_PERSON
                        displayName: Ada Lovelace
                        recordStatus: ARCHIVED
                        createdAt: '2026-09-19T12:00:00.123456Z'
                        version: 3
                    nextCursor: null
                    prevCursor: null
                    totalElements: 1
                    totalPages: 1
                    numberOfElements: 1
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        default:
          $ref: '#/components/responses/DefaultError'
...
  /v1/parties/{partyId}:
    parameters:
      - $ref: '#/components/parameters/PartyIdPath'
      - $ref: '#/components/parameters/TenantId'
      - $ref: '#/components/parameters/ProcessId'
      - $ref: '#/components/parameters/UserId'
    get:
      tags:
        - parties
      operationId: getParty
      summary: Retrieve the common Party detail representation
      description: |
        Returns a tenant-owned Party in DRAFT, ACTIVE, INACTIVE, or ARCHIVED, preserving its stored
        historical text, version, and audit values. No If-Match or Idempotency-Key is required;
        retrieval does not depend on Geographic Reference validation.
        Includes partyId, type, displayName, recordStatus, version, createdAt, updatedAt, createdBy,
        updatedBy, and exactly the matching naturalPersonDetails or legalEntityDetails property.
        Excludes initialIdentifier, identifier collections, nationality collections, tenantId,
        pagination properties, and all protected identifier values, hashes, ciphertext, or keys.
        A noncanonical lowercase UUID returns 400 party-id-invalid. An absent or cross-tenant
        Party returns the same 404 party-not-found without disclosing its owner, type, or version.
        An accepted Process-Id is echoed unchanged on success and failure. Framework 404/405 and
        unexpected 500 failures use the shared status/code-only error envelope.
      responses:
        '200':
          description: Current safe Party details with exactly the immutable type's detail property
          headers:
            Process-Id:
              $ref: '#/components/headers/ProcessIdEcho'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartyDetailApiResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/PartyNotFound'
        default:
          $ref: '#/components/responses/DefaultError'
    patch:
      tags:
        - parties
      operationId: updateParty
      summary: Correct a Party display name
      description: |
        Accepts exactly one supported string property, displayName, in a JSON object. Missing or
        null bodies return 400 request-body-required; missing/null displayName returns
        400 display-name-required. Malformed JSON, non-object bodies, duplicate/unknown properties,
        and incompatible property types return 400 bad-request without scalar coercion.
        Validates trusted context first, then body syntax/required value/normalized maximum length,
        then canonical partyId, then If-Match. Applicable framework checks retain their precedence.
        Normalization uses strip().toUpperCase(Locale.ROOT), retaining accents, punctuation, and
        interior whitespace. More than 300 normalized UTF-16 code units returns
        400 display-name-too-long. An empty or whitespace-only string remains structurally valid
        until business validation and then returns 422 blank-display-name.
        Business failures are evaluated in order: tenant-scoped absence (404 party-not-found),
        expected-version mismatch (412 expected-version-mismatch), then blank-display-name.
        A lost concurrent update also returns 412 expected-version-mismatch.
        A correction is permitted in DRAFT, ACTIVE, INACTIVE, and ARCHIVED. Every accepted write,
        including identical canonical input, advances the Party version exactly once and records
        the requesting User-Id and update time. Identity, type, status, subtype details, independent
        identifiers, nationalities, creation audit values, and registration replay results are preserved.
        Enabled events and the accepted mutation are atomic; a failed attempt leaves no partial outcome.
        Idempotency-Key is not required and supplying it does not enable lifecycle replay for PATCH.
        The response uses the safe root detail representation and echoes the current accepted Process-Id.
      parameters:
        - $ref: '#/components/parameters/PartyIfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PartyUpdateRequest'
      responses:
        '200':
          description: Corrected display name, update audit information, and exactly one version increment
          headers:
            Process-Id:
              $ref: '#/components/headers/ProcessIdEcho'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartyDetailApiResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/PartyNotFound'
        '412':
          $ref: '#/components/responses/PartyExpectedVersionMismatch'
        '422':
          $ref: '#/components/responses/PartyBlankDisplayName'
        default:
          $ref: '#/components/responses/DefaultError'
  /v1/parties/{partyId}/activate:
    parameters:
      - $ref: '#/components/parameters/PartyIdPath'
      - $ref: '#/components/parameters/TenantId'
      - $ref: '#/components/parameters/ProcessId'
      - $ref: '#/components/parameters/UserId'
      - $ref: '#/components/parameters/PartyLifecycleIdempotencyKey'
      - $ref: '#/components/parameters/PartyLifecycleIfMatch'
    post:
      tags:
        - parties
      operationId: activateParty
      summary: Activate an eligible draft Party
      description: |
        Requires no body. After trusted context and applicable framework checks, validates optional
        Idempotency-Key, canonical partyId, then required If-Match, even for a potential saved replay.
        Resolves an equivalent completed keyed result or 409 idempotency-key-conflict before current
        state checks. New-operation failure precedence is tenant-scoped absence (404 party-not-found),
        version mismatch (412 stale-party-version), lifecycle (409 invalid-party-lifecycle), then
        qualifying evidence (422 missing-qualifying-identifier).
        Only DRAFT may become ACTIVE. ACTIVE, INACTIVE, and ARCHIVED reject new activation;
        reactivation is not supported. Requires an identifier owned by the same tenant and Party in
        VERIFIED status, with null expiration or expiration on/after the single UTC evaluation date,
        referencing a matching scheme applicable to this Party type or BOTH. Primary flags and
        DEPRECATED/RETIRED scheme states do not disqualify otherwise eligible evidence. Pending,
        rejected, revoked, expired, unrelated, mismatched, and incompatible evidence cannot qualify.
        Success advances the Party version once and updates audit values while preserving the label,
        subtype details, independent identifier states/versions, and registration replay results.
        The root write, successful keyed result, and enabled activation event are atomic.
        An equivalent tenant/action/key retry with the original Party ID and expected version returns
        the original 200 successful data/version/audit, even after later archival or service restart.
        Equivalent concurrent keyed requests converge on one result instead of returning a stale-version
        loser. Replay does not re-evaluate evidence or produce another mutation/event. Failed attempts
        consume no completed key. The current accepted Process-Id is echoed, while replayed audit values
        retain the original actor/time. Unkeyed or distinct-key operations use current-version arbitration.
      responses:
        '200':
          description: Accepted activation or the identical historical activation result on an equivalent keyed retry
          headers:
            Process-Id:
              $ref: '#/components/headers/ProcessIdEcho'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartyDetailApiResponse'
              example:
                status: 200
                code: successful
                data:
                  partyId: 0198ce2a-7b7d-7ab4-a5cf-4d4d7db89ab1
                  type: NATURAL_PERSON
                  displayName: ADA LOVELACE
                  recordStatus: ACTIVE
                  version: 1
                  createdAt: '2026-09-19T12:00:00Z'
                  updatedAt: '2026-09-19T12:01:00Z'
                  createdBy: registrar
                  updatedBy: activator
                  naturalPersonDetails: { givenNames: ADA, familyNames: LOVELACE }
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/PartyNotFound'
        '409':
          $ref: '#/components/responses/PartyLifecycleConflict'
        '412':
          $ref: '#/components/responses/PartyStaleVersion'
        '422':
          $ref: '#/components/responses/PartyMissingActivationEvidence'
        default:
          $ref: '#/components/responses/DefaultError'
  /v1/parties/{partyId}/deactivate:
    parameters:
      - $ref: '#/components/parameters/PartyIdPath'
      - $ref: '#/components/parameters/TenantId'
      - $ref: '#/components/parameters/ProcessId'
      - $ref: '#/components/parameters/UserId'
      - $ref: '#/components/parameters/PartyLifecycleIdempotencyKey'
      - $ref: '#/components/parameters/PartyLifecycleIfMatch'
    post:
      tags:
        - parties
      operationId: deactivateParty
      summary: Deactivate an active Party
      description: |
        Requires no body. After trusted context and applicable framework checks, validates optional
        Idempotency-Key, canonical partyId, then required If-Match, including requests for saved results.
        Resolves completed replay or idempotency-key-conflict before tenant-scoped absence, stale
        version, then lifecycle validation, using 404 party-not-found, 412 stale-party-version,
        and 409 invalid-party-lifecycle respectively. Only ACTIVE may become INACTIVE; DRAFT,
        INACTIVE, and ARCHIVED reject new deactivation. No identifier evidence is required.
        Success increments the Party version exactly once and updates audit values, preserving all
        identity, label, subtype details, independent identifiers, nationalities, creation audit,
        and registration replay results. Root changes, the successful keyed result, and enabled
        deactivation events are atomic; rejection leaves no partial outcome or completed key.
        Equivalent tenant/action/key retries return the original 200 successful Party data, version,
        and audit values after subsequent changes or restart, without another mutation/event. Concurrent
        equivalent keyed retries converge. Current User-Id and Process-Id do not change equivalence;
        the current accepted Process-Id is echoed while the saved audit actor/time remain historical.
        Distinct-key or unkeyed requests losing the version race return 412 stale-party-version.
      responses:
        '200':
          description: Accepted deactivation or its historical successful keyed result
          headers:
            Process-Id:
              $ref: '#/components/headers/ProcessIdEcho'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartyDetailApiResponse'
              example:
                status: 200
                code: successful
                data:
                  partyId: 0198ce2b-d6a3-7d6e-80ba-d97b21d793e5
                  type: LEGAL_ENTITY
                  displayName: ANALYTICAL ENGINES
                  recordStatus: INACTIVE
                  version: 2
                  createdAt: '2026-09-19T12:00:00Z'
                  updatedAt: '2026-09-19T12:02:00Z'
                  createdBy: registrar
                  updatedBy: operator
                  legalEntityDetails: { legalName: ANALYTICAL ENGINES, incorporationCountryCode: GB }
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/PartyNotFound'
        '409':
          $ref: '#/components/responses/PartyLifecycleConflict'
        '412':
          $ref: '#/components/responses/PartyStaleVersion'
        default:
          $ref: '#/components/responses/DefaultError'
  /v1/parties/{partyId}/archive:
    parameters:
      - $ref: '#/components/parameters/PartyIdPath'
      - $ref: '#/components/parameters/TenantId'
      - $ref: '#/components/parameters/ProcessId'
      - $ref: '#/components/parameters/UserId'
      - $ref: '#/components/parameters/PartyLifecycleIdempotencyKey'
      - $ref: '#/components/parameters/PartyLifecycleIfMatch'
    post:
      tags:
        - parties
      operationId: archiveParty
      summary: Archive a retained Party
      description: |
        Requires no body. After trusted context and applicable framework checks, validates optional
        Idempotency-Key, canonical partyId, then required If-Match, including requests for saved results.
        Resolves completed replay or idempotency-key-conflict before tenant-scoped absence, stale
        version, then lifecycle validation, using 404 party-not-found, 412 stale-party-version,
        and 409 invalid-party-lifecycle respectively. DRAFT, ACTIVE, and INACTIVE may become ARCHIVED
        directly, with no identifier evidence requirement. ARCHIVED rejects every new lifecycle action
        at its current version; archival does not prevent a later descriptive root PATCH.
        Success increments the Party version exactly once and updates audit values without physically
        deleting records. Identity, label, subtype details, independent identifiers and their versions,
        nationalities, creation audit, and registration replay results are retained. Root changes,
        the successful keyed result, and enabled archival events are atomic; rejection leaves no
        partial outcome or completed key. Archived Parties remain retrievable and eligible for listing.
        Equivalent tenant/action/key retries return the original 200 successful Party data, version,
        and audit after later corrections or restart. A lost response after commit does not undo
        acceptance. Replay never changes current state or emits another event; equivalent concurrent
        retries converge. The current accepted Process-Id is echoed while audit values remain historical.
        Distinct-key or unkeyed requests losing the version race return 412 stale-party-version.
      responses:
        '200':
          description: Accepted archival or its historical successful keyed result
          headers:
            Process-Id:
              $ref: '#/components/headers/ProcessIdEcho'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartyDetailApiResponse'
              example:
                status: 200
                code: successful
                data:
                  partyId: 0198ce2a-7b7d-7ab4-a5cf-4d4d7db89ab1
                  type: NATURAL_PERSON
                  displayName: ADA LOVELACE
                  recordStatus: ARCHIVED
                  version: 3
                  createdAt: '2026-09-19T12:00:00Z'
                  updatedAt: '2026-09-19T12:03:00Z'
                  createdBy: registrar
                  updatedBy: archivist
                  naturalPersonDetails: { givenNames: ADA, familyNames: LOVELACE }
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/PartyNotFound'
        '409':
          $ref: '#/components/responses/PartyLifecycleConflict'
        '412':
          $ref: '#/components/responses/PartyStaleVersion'
        default:
          $ref: '#/components/responses/DefaultError'
...
components:
  parameters:
    TenantId:
      name: Tenant-Id
      in: header
      required: true
      description: >-
        Trusted tenant identifier. Exactly one canonical UUID value is required.
        Failures return `tenant-id-required`, `tenant-id-duplicated`, or `tenant-id-invalid`.
      schema:
        type: string
        format: uuid
        pattern: '^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$'
    ProcessId:
      name: Process-Id
      in: header
      required: true
      description: >-
        Trusted process identifier. Exactly one canonical UUID value is required. An accepted value
        is echoed unchanged in the response.
        Failures return `process-id-required`, `process-id-duplicated`, or `process-id-invalid`.
      schema:
        type: string
        format: uuid
        pattern: '^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$'
    UserId:
      name: User-Id
      in: header
      required: true
      description: >-
        Trusted audit subject. Exactly one non-blank value without control characters is required.
        The limit is 128 Unicode code points. C0, DEL, and C1 control characters are prohibited.
        Failures return `user-id-required`, `user-id-duplicated`, `user-id-blank`,
        `user-id-too-long`, or `user-id-unsafe`.
      schema:
        type: string
        minLength: 1
        maxLength: 128
        pattern: '^(?=.*\S)[^\x00-\x1F\x7F-\x9F]+$'
...
    PartyIdPath:
      name: partyId
      in: path
      required: true
      description: Canonical lowercase Party UUID; invalid syntax returns 400 party-id-invalid.
      schema:
        type: string
        format: uuid
        pattern: '^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$'
...
    PartyCursorQuery:
      name: cursor
      in: query
      required: false
      description: >-
        One opaque server-issued continuation scoped to tenant, effective filters, and page size.
        Omit for the first page. Empty, malformed, altered, or scope-mismatched values return
        400 bad-request. Equivalent normalized filters and equivalent offsets preserve scope.
      schema:
        type: string
        minLength: 1
    PartyLimitQuery:
      name: limit
      in: query
      required: false
      description: >-
        One decimal integer from 1 through 200; defaults to 50 only when absent. Blank, fractional,
        negative, zero, or oversized values return 400 bad-request. Included in cursor scope.
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
...
  responses:
...
    Unauthorized:
      description: Falta autenticación o credenciales inválidas.
      headers:
        Process-Id:
          $ref: '#/components/headers/ProcessIdEcho'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          example:
            status: 401
            code: unauthorized
...
    DefaultError:
      description: Error inesperado.
      headers:
        Process-Id:
          $ref: '#/components/headers/ProcessIdEcho'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          example:
            status: 500
            code: server-error
  headers:
    ProcessIdEcho:
      description: >-
        Unchanged accepted `Process-Id`. This header is omitted when the request value is missing,
        duplicated, or invalid.
      schema:
        type: string
        format: uuid
        pattern: '^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$'
  schemas:
...
    PartyType:
      type: string
      enum:
        - NATURAL_PERSON
        - LEGAL_ENTITY
    PartyRecordStatus:
      type: string
      enum:
        - DRAFT
        - ACTIVE
        - INACTIVE
        - ARCHIVED
...